# specs/002_api_specification.yaml
...
openapi: "3.1.0"
info:
  title: Cyber Chakra Forensics Admin Portal API
  version: "1.0.0"
  description: |
    REST API for the CMF Admin Portal and the public-facing desktop app endpoints.
    All admin endpoints require Bearer token authentication.
    All public endpoints require a valid license key (X-License-Key header).
servers:
  - url: https://cyberchakra.online
    description: Production
  - url: https://staging.cyberchakra.online
    description: Staging

# =============================================================================
# Security Schemes
# =============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin portal JWT token obtained from /api/v1/auth/login.

    LicenseKeyAuth:
      type: apiKey
      in: header
      name: X-License-Key
      description: License key for public desktop app endpoints.

  # ===========================================================================
  # Shared Schemas
  # ===========================================================================
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 25
        total:
          type: integer
          example: 142
        total_pages:
          type: integer
          example: 6
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
    Announcement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        body_html:
          type: string
        announcement_type:
          type: string
          enum: [info, warning, critical, maintenance, feature, promotion]
        target:
          type: string
          enum: [all, trial, standard, professional, enterprise, government, specific_orgs]
        is_dismissible:
          type: boolean
        is_pinned:
          type: boolean
        action_url:
          type: string
        action_label:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        is_published:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ---- Support Ticket ----
...
    PublicAnnouncementItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body_html:
          type: string
        announcement_type:
          type: string
        is_dismissible:
          type: boolean
        is_pinned:
          type: boolean
        action_url:
          type: string
        action_label:
          type: string
        starts_at:
          type: string
          format: date-time

  # ===========================================================================
  # Shared Parameters
  # ===========================================================================
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QueryPage:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    QueryPerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/announcements:
    get:
      tags: [Announcements]
      summary: List announcements
      parameters:
        - $ref: '#/components/parameters/QueryPage'
        - $ref: '#/components/parameters/QueryPerPage'
        - name: is_published
          in: query
          schema:
            type: boolean
        - name: announcement_type
          in: query
          schema:
            type: string
        - name: active_only
          in: query
          schema:
            type: boolean
            description: "Only return currently active (started, not ended)"
      responses:
        "200":
          description: Announcements list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Announcement'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
    post:
      tags: [Announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body, announcement_type]
              properties:
                title:
                  type: string
                body:
                  type: string
                  description: "Markdown content"
                announcement_type:
                  type: string
                  enum: [info, warning, critical, maintenance, feature, promotion]
                target:
                  type: string
                  enum: [all, trial, standard, professional, enterprise, government, specific_orgs]
                  default: all
                target_org_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                target_versions:
                  type: array
                  items:
                    type: string
                target_platforms:
                  type: array
                  items:
                    type: string
                min_version:
                  type: string
                max_version:
                  type: string
                is_dismissible:
                  type: boolean
                  default: true
                is_pinned:
                  type: boolean
                  default: false
                action_url:
                  type: string
                action_label:
                  type: string
                starts_at:
                  type: string
                  format: date-time
                ends_at:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Announcement created (unpublished)
  /api/v1/announcements/{id}:
    get:
      tags: [Announcements]
      summary: Get announcement
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        "200":
          description: Announcement details
    patch:
      tags: [Announcements]
      summary: Update announcement
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                announcement_type:
                  type: string
                target:
                  type: string
                target_org_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                is_dismissible:
                  type: boolean
                is_pinned:
                  type: boolean
                action_url:
                  type: string
                action_label:
                  type: string
                starts_at:
                  type: string
                  format: date-time
                ends_at:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Announcement updated
    delete:
      tags: [Announcements]
      summary: Delete announcement
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        "204":
          description: Announcement deleted
...
  /api/public/v1/announcements:
    get:
      tags: [Public - Content]
      summary: Get active announcements for desktop app
      security: []
      parameters:
        - name: app_version
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
        - name: license_key
          in: query
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: string
            format: date-time
            description: "Only return announcements created/updated after this time"
      responses:
        "200":
          description: Active announcements
          content:
            application/json:
              schema:
                type: object
                properties:
                  announcements:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicAnnouncementItem'
...